+++ Braintree.Tests/MerchantAccount.cs
@@ -7,5 +7,6 @@ namespace Braintree.Tests
     {
         public static String DEFAULT_MERCHANT_ACCOUNT_ID = "sandbox_credit_card";
         public static String NON_DEFAULT_MERCHANT_ACCOUNT_ID = "sandbox_credit_card_non_default";
+        public static String NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID = "sandbox_sub_merchant_account";
     }
 }