+++ Braintree/Transaction.cs
@@ -127,6 +127,7 @@ namespace Braintree
         public TransactionType Type { get; protected set; }
         public DateTime? UpdatedAt { get; protected set; }
         public Dictionary<String, String> CustomFields { get; protected set; }
+        public DisbursementDetails DisbursementDetails { get; protected set; }
 
         private BraintreeService Service;
 
@@ -183,6 +184,7 @@ namespace Braintree
             CurrencyIsoCode = node.GetString("currency-iso-code");
             CvvResponseCode = node.GetString("cvv-response-code");
             Descriptor = new Descriptor(node.GetNode("descriptor"));
+            DisbursementDetails = new DisbursementDetails(node.GetNode("disbursement-details"));
 
             BillingAddress = new Address(node.GetNode("billing"));
             ShippingAddress = new Address(node.GetNode("shipping"));
@@ -315,6 +317,11 @@ namespace Braintree
             if (ShippingAddress.Id == null) return null;
 
             return new AddressGateway(Service).Find(Customer.Id, ShippingAddress.Id);
+        }
+
+        public Boolean IsDisbursed()
+        {
+          return DisbursementDetails.IsValid();
         }
     }
 }